release cleanup
- forgot to commit API update
- lint fix for toolbar and DRYer
- move editor types
- bump `platform-editor` version
