+++ packages/platform/src/editor/editor.model.ts
@@ -1,4 +1,93 @@
-import { TextDirection, UsjNodeOptions, ViewOptions } from "shared-react";
+import { Usj } from "@eten-tech-foundation/scripture-utilities";
+import { SerializedVerseRef } from "@sillsdev/scripture";
+import { RefObject } from "react";
+import { LoggerBasic } from "shared";
+import {
+  AnnotationRange,
+  DeltaOp,
+  DeltaSource,
+  SelectionRange,
+  TextDirection,
+  UsjNodeOptions,
+  ViewOptions,
+} from "shared-react";
+
+/**
+ * Forward reference for the editor.
+ *
+ * @public
+ */
+
+export interface EditorRef {
+  /** Focus the editor. */
+  focus(): void;
+  /** Get USJ Scripture data. */
+  getUsj(): Usj | undefined;
+  /** Set the USJ Scripture data. */
+  setUsj(usj: Usj): void;
+  /** EXPERIMENTAL: Apply Operational Transform delta update */
+  applyUpdate(ops: DeltaOp[], source?: DeltaSource): void;
+  /**
+   * Get the selection location or range.
+   * @returns the selection location or range, or `undefined` if there is no selection. The
+   *   json-path in the selection assumes no comment Milestone nodes are present in the USJ.
+   */
+  getSelection(): SelectionRange | undefined;
+  /**
+   * Set the selection location or range.
+   * @param selection - A selection location or range. The json-path in the selection assumes no
+   *   comment Milestone nodes are present in the USJ.
+   */
+  setSelection(selection: SelectionRange): void;
+  /**
+   * Add an ephemeral annotation.
+   * @param selection - An annotation range containing the start and end location. The json-path in
+   *   an annotation location assumes no comment Milestone nodes are present in the USJ.
+   * @param type - Type of the annotation.
+   * @param id - ID of the annotation.
+   */
+  addAnnotation(selection: AnnotationRange, type: string, id: string): void;
+  /**
+   * Remove an ephemeral annotation.
+   * @param type - Type of the annotation.
+   * @param id - ID of the annotation.
+   */
+  removeAnnotation(type: string, id: string): void;
+  /**
+   * Insert a note at the specified selection, e.g. footnote, cross-reference, endnote.
+   * @param marker - The marker type for the note.
+   * @param caller - Optional note caller to override the default for the given marker.
+   * @param selection - Optional selection range where the note should be inserted. By default it
+   *   will use the current selection in the editor.
+   * @throws Will throw an error if the marker is not a valid note marker.
+   */
+  insertNote(marker: string, caller?: string, selection?: SelectionRange): void;
+  /** Ref to the end of the toolbar - INTERNAL USE ONLY to dynamically add controls in the toolbar. */
+  toolbarEndRef: RefObject<HTMLElement | null> | null;
+}
+
+/**
+ * Props for the Editor component that provides Scripture editing functionality.
+ *
+ * @public
+ */
+
+export interface EditorProps<TLogger extends LoggerBasic> {
+  /** Initial Scripture data in USJ format. */
+  defaultUsj?: Usj;
+  /** Scripture reference that controls the general cursor location of the Scripture. */
+  scrRef?: SerializedVerseRef;
+  /** Callback function when the Scripture reference has changed. */
+  onScrRefChange?: (scrRef: SerializedVerseRef) => void;
+  /** Callback function when the cursor selection changes. */
+  onSelectionChange?: (selection: SelectionRange | undefined) => void;
+  /** Callback function when USJ Scripture data has changed. */
+  onUsjChange?: (usj: Usj, ops?: DeltaOp[], source?: DeltaSource) => void;
+  /** Options to configure the editor. */
+  options?: EditorOptions;
+  /** Logger instance. */
+  logger?: TLogger;
+}
 
 /**
  * Options to configure the editor.
