+++ packages/platform/src/index.ts
@@ -41,8 +41,7 @@ export {
   viewModeToViewNames,
 } from "shared-react";
 
-export type { EditorProps, EditorRef } from "./editor/Editor";
-export type { EditorOptions } from "./editor/editor.model";
+export type { EditorOptions, EditorProps, EditorRef } from "./editor/editor.model";
 export type { CommentBase, Comments, Thread } from "./marginal/comments/commenting";
 export type { MarginalRef, MarginalProps } from "./marginal/Marginal";
 export type { LoggerBasic, NodeOptions } from "shared";
